Custom icon chooser implemented
* IconsHandler retrieves icons from IconPackXML or SystemIconPack
* UserHandle is now Parcelable

+++ app/src/main/java/fr/neamar/kiss/db/AppRecord.java
@@ -4,7 +4,7 @@ public class AppRecord {
     public static final int FLAG_CUSTOM_NAME = 2;
     public static final int FLAG_CUSTOM_ICON = 4;
 
-    public int dbId;
+    public long dbId;
 
     public String name;
